pyproject.toml
[project]
name = "learn-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.1",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.19.0",
]
